LayoutTests/http/tests/security/mixedContent/insecure-iframe-in-main-frame.html

<html>
<body>
<script>
if (window.layoutTestController) {
    layoutTestController.waitUntilDone();
    layoutTestController.dumpAsText();
    layoutTestController.dumpFrameLoadCallbacks();
    layoutTestController.setCanOpenWindows();
    layoutTestController.setCloseRemainingWindowsWhenComplete(true);
}

window.addEventListener("message", function (e) {
  if (window.layoutTestController)
      layoutTestController.notifyDone();
}, false);

</script>
<p>This test opens a window that loads an insecure iframe.  We should trigger a
mixed content callback because the main frame in the window is HTTPS but is
displaying insecure content.</p>
<script>
window.open("https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-frame.html");
</script>
</body>
</html>
